css/css-scroll-snap-2/snapchanged/snapchanged-same-targets-after-layout-changed.html

<!DOCTYPE html>
<html>

<head>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#snap-events" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/dom/events/scrolling/scroll_support.js"></script>
  <style>
    #scroller {
      overflow-y: scroll;
      scroll-snap-type: y mandatory;
      width: 500px;
      height: 500px;
      background-color: white;
      position: relative;
    }
    .space_filler {
      display: inline-block;
      width: 40%;
      height: 30%;
      background-color: green;
    }
    .snap_area {
      scroll-snap-align: start;
      background-color: yellow;
      position: absolute;
      width: 40%;
      height: 30%;
    }

    .left {
      left: 1px;
    }
    .right {
      left: 41%;
    }
  </style>
</head>
<body>
  <div id="scroller">
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div class="space_filler"></div>
    <div id="left" class="snap_area left"><h1>1</h1></div>
    <div id="right" class="snap_area right"><h1>2</h1></div>
  </div>
  <script>
    function switch_boxes() {
      right.style.left = right.style.left == "1px" ? "41%" : "1px";
      left.style.left = left.style.left == "41%" ? "1px" : "41%";
    }

    promise_test(async (t) => {
      await waitForCompositorCommit();
      scroller.onsnapchanged = t.unreached_func("snapchanged shouldn't fire " +
          "since the scroller is snapped to the same elements.");
      assert_greater_than(right.offsetLeft, left.offsetLeft,
          "boxes have switched positions.");
      switch_boxes();
      await waitForCompositorCommit();
      assert_less_than(right.offsetLeft, left.offsetLeft,
          "boxes have switched positions.");
      await waitForCompositorCommit();
    }, "snapchanged doesn't fire after layout change if snapped to the same " +
    "elements");

  </script>
</body>
</html>
